Remove log-error mysqld_safe option
According to [1], if file name has no extension,
the server adds an extension of ".err".

It causes two different issues.

The first one is impossibility to create /dev/stdout.err file
in newer versions of mysql (5.6.35 at least).

The second one is a potential OOM in case of intensive stderr flow from mysql.
Because the /dev mount point is a devtmpfs which is in-memory pseudo file system.

[1] https://dev.mysql.com/doc/refman/5.6/en/server-options.html#option_mysqld_log-error

Related-Bug: #1657707

Change-Id: I430d8dd085433e8cb84fdf0595b8357758d9ea75