Add missing log level argument to ocf_log calls
There are a few ocf_log calls introduced recently in the mysql-wss OCF
file to ease troubleshooting and debugging. But, due to the absence of
the log level argument, error entries, indicating the missing argument,
are started to appear in the logs.

As a solution, 'debug' log level is added to all these discriminated
ocf_log call occurances.

Closes-bug: 1666507
Closes-Bug: 1683647

Change-Id: I59fe68cf53d361d565e743dc10b014e122139cb3
Signed-off-by: Gabor Orosz <gabor.orosz@ericsson.com>