Correct logical expressions for storage backends
After introducing the change Iead5167210c4132badb866afc25d4ef14e27f6b2
swift isn't installed if ceph used as a backend for images
and not for object-storage service. In this sutiation we end up
with absence of object-storage at all which is wrong.

That commit changes logic. If we have RadosGW, then we have no Swift.
If we have no RadosGW, then we have Swift both for images and for
object store.

Change-Id: If319845aa9f2d9b3ca07936350ef374f504679cb
Closes-Bug: #1604879