Introduce critical section on master election process
This commit adds a global cluster attribute with value which
means the timestamp of when the election process started.
If we have election on any node in process we sleep for a while
unless the attribute is outdated.

We start the election only if the attribute is outdated or if it
does not exist.

This prevents us from hitting rare condition when several nodes
start simultaneously but do not agree on the master node due to
race condition in MySQL start time and pacemaker attribute setting

Change-Id: I7f4728b75ce5577338dff182634b608823cff74e
Closes-bug: #1617400